First implementation of error sample weighting. It may be decaying too fast.

+++ src/iterator.h
@@ -1,17 +1,19 @@
 #ifndef ITERATOR_H_INCL
 #define ITERATOR_H_INCL
 
+#include <sys/types.h>
+
 class Iterator {
   public:
     virtual double probability() = 0;
-    virtual double probability(int pos) = 0;
+    virtual double probability(size_t pos) = 0;
     virtual double value() = 0;
     virtual void advance() = 0;
     virtual bool isDone() = 0;
     virtual void reset() = 0;
     virtual int position() = 0;
     virtual int totalCount() = 0;
-    virtual double at(int pos) = 0;
+    virtual double at(size_t pos) = 0;
     virtual ~Iterator() {}
 };
